backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "regression-platform-backend"
version = "0.1.0"
description = "다중 턴 테이블형 회귀 분석 플랫폼 백엔드"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    # Cache / Queue
    "redis>=5.0.4",
    "rq>=1.16.2",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # LangGraph / LangChain
    "langgraph>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",
    # Data / ML
    "pandas>=2.2.0",
    "pyarrow>=16.0.0",
    "lightgbm>=4.3.0",
    "shap>=0.45.0",
    "optuna>=3.6.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.20.0",
    "statsmodels>=0.14.0",
    "xgboost>=2.0.0",
    "catboost>=1.2.0",
    "ydata-profiling>=4.6.0",
    "openpyxl>=3.1.2",
    "numpy>=1.26.0",
    # Settings / Logging
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    # Utilities
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
    "python-dateutil>=2.9.0",
    # PandasAI
    "pandasai>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
